+++ changelog.md
@@ -1,13 +1,14 @@
 # Changelog
 
 * Break vanilla fletching + message on right click
-* Curse of Unhurried pants only?
 
 ## Upcoming
 * Angry Creepers now use the EAI default sound
 * Sea creatures size adjusted (0.4x\~1.75x -> 0.5x\~1.6x)
 * Moved Serene Seasons integration in IguanaTweaks Reborn
 * Curse of Unhurried is now appliable only to leggings
+* Vanilla Fletching tables now show an error when right-clicked and will drop ITE's one when broken
+* Fixed Vanilla blast furnaces showing the "invalid block" message 4 times
 
 ## 1.12.4
 * Rebalanced Armor Piercer